Add test script
Uses tape for now. Jest is not supported out of the box, though it is possible to set it up
