+++ package.json
@@ -9,8 +9,8 @@
   },
   "scripts": {
     "lint": "eslint . --max-warnings 0",
-    "check": "yarn run lint",
-    "test": "echo 'no tests, yet'"
+    "check": "yarn lint && yarn test",
+    "test": "tape plugins/**/*.test.js"
   },
   "engines": {
     "node": ">=8.9.0"
@@ -41,7 +41,8 @@
   "devDependencies": {
     "eslint": "^4.19.1",
     "eslint-plugin-shopify": "^20.0.0",
-    "stylelint-test-rule-tape": "^0.2.0"
+    "stylelint-test-rule-tape": "^0.2.0",
+    "tape": "^4.9.1"
   },
   "peerDependencies": {
     "stylelint": ">=9.0"
